+++ jvm-streaming/src/test/haskell/Main.hs
@@ -1,11 +1,16 @@
-{-# LANGUAGE OverloadedStrings #-}
 module Main where
 
+import Data.List (intersperse)
+import qualified Data.Text as Text
+import qualified Data.Text.Encoding as Text
 import Language.Java (withJVM)
+import Language.Java.Streaming.Jars (getJars)
 import qualified Spec
 import Test.Hspec
 
 main :: IO ()
-main =
-    withJVM ["-Djava.class.path=../jvm-batching/build/libs/jvm-batching.jar"] $
+main = do
+    jars <- getJars
+    let classpath = concat $ intersperse ":" jars
+    withJVM [Text.encodeUtf8 $ Text.pack $ "-Djava.class.path=" ++ classpath] $
       hspec Spec.spec